third_party/blink/web_tests/fast/dom/shadow/touch-event-expected.txt

Touch event should be fired in Shadow DOM.

Elements in ShadowDOM should not be revealed in touchTarget if it's examined in non shadow tree.

PASS touchstartFiredInShadowDOM is true
PASS touchTargetForHost1 is host1
PASS touchTargetForInput1 is input1
PASS successfullyParsed is true

TEST COMPLETE
